Fix a segfault
I don't know how I got it, but I ended up with the line
 57 1
in my location history file. when you did the strtok it did not find enough tokens and one of them ended up returning NULL which atoi dereferenced segfaulting the program. This fix may not fix the actual problem of how that actually ended up in the file, but the program does not segfault  anymore.